health_check/pyproject.toml
[tool.poetry]
name = "health-monitor"
version = "0.1.0"
description = "Serviço de health-check com métricas Prometheus"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
psycopg2-binary = "^2.9.9"
redis = "^5.0.1"
kubernetes = "^34.0.0"
requests = "^2.31.0"
boto3 = "^1.34.0"
openai = "^1.12.0"
prometheus-client = "^0.19.0"
aiohttp = "^3.9.1"
python-dotenv = "^1.2.0"

[tool.mypy]
ignore_missing_imports = true
explicit_package_bases = true 
namespace_packages = true
mypy_path = "."

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"